demo_project.requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in demo_project.requirements.in, then run this to update:
#
#    pip-compile demo_project.requirements.in
#
django-classy-tags==0.6.2
django-model-utils==2.3.1
django-template-finder==0.6.2
django==1.8.5
djangorestframework==3.3.0
